requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
PyPDF2==3.0.1
python-docx==0.8.11
flask==2.3.2
requests==2.31.0
google-generativeai==0.3.0
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.25.0